remove call to initialize() on the token impl in cloneable tests
