+++ solidity/scripts/deploy_cloneable.ts
@@ -1,7 +1,10 @@
 import { ethers, ignition } from "hardhat";
 import { Logger, ILogObj } from "tslog";
 const logLevel = process.env.LOG_LEVEL || "3";
-export const logger: Logger<ILogObj> = new Logger({ name: "deploy_cloneable", minLevel: parseInt(logLevel) });
+export const logger: Logger<ILogObj> = new Logger({
+  name: "deploy_cloneable",
+  minLevel: parseInt(logLevel),
+});
 
 import erc20Module from "../ignition/modules/erc20";
 import { getLinkedContractFactory, deploy } from "./lib/common";
@@ -21,14 +24,15 @@ export async function deployFungible(tokenName: string) {
 
   const zetoImpl: any = await zetoFactory.deploy();
   await zetoImpl.waitForDeployment();
-  // console.log(args);
-  await zetoImpl.connect(deployer).initialize(...args);
-
-  const tx3 = await zetoImpl.connect(deployer).setERC20(erc20.target);
-  await tx3.wait();
+  // Do not call {initialize} or {setERC20} on the implementation. Leaf Zeto
+  // contracts lock the impl with {_disableInitializers} in the constructor,
+  // so {initialize} here would always revert with {InvalidInitialization}.
+  // Factory tests ({ZetoTokenFactory}) deploy ERC1967 proxies whose constructor
+  // delegates {initialize} into fresh proxy storage; {deployZeto} then binds
+  // ERC20 on each proxy via {setERC20}.
 
   logger.debug(`ERC20 deployed:     ${erc20.target}`);
-  logger.debug(`ZetoToken deployed: ${zetoImpl.target}`);
+  logger.debug(`ZetoToken impl deployed: ${zetoImpl.target}`);
 
   return { deployer, zetoImpl, erc20, args };
 }
@@ -46,9 +50,10 @@ export async function deployNonFungible(tokenName: string) {
   }
   const zetoImpl: any = await zetoFactory.deploy();
   await zetoImpl.waitForDeployment();
-  await zetoImpl.connect(deployer).initialize(...args);
+  // Same rationale as {deployFungible}: impl stays uninitialized; proxies
+  // created by the factory run {initialize} via ERC1967Proxy constructor data.
 
-  logger.debug(`ZetoToken deployed: ${zetoImpl.target}`);
+  logger.debug(`ZetoToken impl deployed: ${zetoImpl.target}`);
 
   return { deployer, zetoImpl, args };
 }
